requirements.txt
# Core
numpy>=1.23
pandas>=2.0
scikit-learn>=1.3

# NLP
spacy>=3.7,<3.8
pydantic<2
sentence-transformers>=2.6
transformers>=4.40
torch

# Knowledge graph / viz
networkx>=3.1
matplotlib>=3.7

# Data fetching / utils
requests>=2.31
beautifulsoup4>=4.12
python-dotenv>=1.0
newsapi-python>=0.2